Adicao de arquivo

+++ resources/views/funcionarios/gerenciar-funcionario.blade.php
@@ -14,7 +14,7 @@
                             <input class="form-control" maxlength="9" type="text" id="2" name="idt" value="{{$idt}}">
                         </div>
                         <div class="col-3">Nome
-                            <input class="form-control" maxlength="50" type="text" id="3" name="nome" value="{{$nome}}">
+                            <input class="form-control" maxlength="50" type="text" id="3" name="nome"value="{{$nome}}">
                         </div>
                         <div class="col-2">Situação
                             <select class="form-select" id="4" name="status" type="number" required="required">
@@ -24,19 +24,11 @@
                             </select>
                         </div>
                         <div class="col"><br>
-                            <input class="btn btn-light btn-sm"
-                                   style="font-size: 0.9rem; box-shadow: 1px 2px 5px #000000; margin:5px;" type="submit"
-                                   value="Pesquisar">
+                            <input class="btn btn-light btn-sm" style="font-size: 0.9rem; box-shadow: 1px 2px 5px #000000; margin:5px;" type="submit" value="Pesquisar">
 
-                            <a href="/gerenciar-funcionario">
-                                <button class="btn btn-light btn-sm " type="button" value=""
-                                        style="font-size: 0.9rem; box-shadow: 1px 2px 5px #000000; margin:5px;">Limpar
-                                </button>
-                            </a>
+                            <a href="/gerenciar-funcionario"><button class="btn btn-light btn-sm " type="button" value="" style="font-size: 0.9rem; box-shadow: 1px 2px 5px #000000; margin:5px;" >Limpar</button></a>
                 </form>
-                <a href="/informar-dados"><input class="btn btn-success btn-sm" type="button" name="6"
-                                                 value="Novo Cadastro +"
-                                                 style="font-size: 0.9rem; box-shadow: 1px 2px 5px #000000; margin:5px;"></a>
+                <a href="/informar-dados"><input class="btn btn-success btn-sm" type="button" name="6" value="Novo Cadastro +" style="font-size: 0.9rem; box-shadow: 1px 2px 5px #000000; margin:5px;"></a>
 
             </div>
         </div>
@@ -65,49 +57,17 @@
                         <td scope="" style="text-align: center;">Inativo</td>
                     @endif
                     <td scope="">
-                        <a href="/editar-funcionario/{{$listas->idf}}">
-                            <button type="button" class="btn btn-outline-warning btn-sm" data-bs-toggle="tooltip"
-                                    data-bs-placement="top" data-bs-title="Editar Dados"><i class="bi-pencil"
-                                                                                            style="font-size: 1rem; color:#303030;"></i>
-                            </button>
-                        </a>
-                        <a href="/pessoas-funcionario/{{$listas->idf}}">
-                            <button type="button" class="btn btn-outline-primary btn-sm"><i class="bi-search"
-                                                                                            style="font-size: 1rem; color:#303030;"></i>
-                            </button>
-                        </a>
-                        <a href="/gerenciar-dependentes/{{$listas->idf}}">
-                            <button type="button" class="btn btn-outline-primary btn-sm"><i class="bi-people-fill"
-                                                                                            style="font-size: 1rem;color:#303030; "></i>
-                            </button>
-                        </a>
-                        <a href="/gerenciar-dados-bancarios/{{$listas->idf}}">
-                            <button type="button" class="btn btn-outline-primary btn-sm"><i class="bi bi-bank2"
-                                                                                            style="font-size: 1rem; color:#303030;"></i>
-                            </button>
-                        </a>
-                        <a href="/editar-funcionario/{{$listas->idf}}">
-                            <button type="button" class="btn btn-outline-primary btn-sm"><i class="bi-bandaid"
-                                                                                            style="font-size: 1rem;color:#303030;"></i>
-                            </button>
-                        </a>
-                        <a href="/gerenciar-certificados/{{$listas->idf}}">
-                            <button type="button" class="btn btn-outline-primary btn-sm"><i class="bi-mortarboard"
-                                                                                            style="font-size: 1rem; color:#303030;"></i>
-                            </button>
-                        </a>
-                        <a href="/gerenciar-acordos/{{$listas->idf}}">
-                            <button type="button" class="btn btn-outline-primary btn-sm"><i class="bi bi-pencil-square"
-                                                                                            style="font-size: 1rem; color:#303030;"></i>
-                            </button>
-                        </a>
+                        <a href="/editar-funcionario/{{$listas->idf}}"><button type="button" class="btn btn-outline-warning btn-sm"><i class="bi-pencil" style="font-size: 1rem; color:#303030;"></i></button></a>
+
+                        <a href="/pessoas-funcionario/{{$listas->idf}}"><button type="button" class="btn btn-outline-primary btn-sm"><i class="bi-search" style="font-size: 1rem; color:#303030;"></i></button></a>
+                        <a href="/gerenciar-dependentes/{{$listas->idf}}"><button type="button" class="btn btn-outline-primary btn-sm"><i class="bi-people-fill" style="font-size: 1rem;color:#303030; "></i></button></a>
+                        <a href="/gerenciar-dados-bancarios/{{$listas->idf}}"><button type="button" class="btn btn-outline-primary btn-sm"><i class="bi bi-bank" style="font-size: 1rem; color:#303030;"></i></button></a>
+                        <a href="/editar-funcionario/{{$listas->idf}}"><button type="button" class="btn btn-outline-primary btn-sm"><i class="bi-bandaid" style="font-size: 1rem;color:#303030;"></i></button></a>
+                        <a href="/gerenciar-certificados/{{$listas->idf}}"><button type="button" class="btn btn-outline-primary btn-sm"><i class="bi-mortarboard" style="font-size: 1rem; color:#303030;"></i></button></a>
+                        <a href="/gerenciar-acordos/{{$listas->idf}}"><button type="button" class="btn btn-outline-primary btn-sm"><i class="bi bi-pencil-square" style="font-size: 1rem; color:#303030;"></i></button></a>
 
 
-                        <button type="button" class="btn btn-outline-danger btn-sm" data-bs-toggle="modal"
-                                data-bs-target="#A{{ $listas->cpf}}-{{ $listas->idf }}"
-                                class="btn btn-outline-danger btn-sm"><i class="bi-trash"
-                                                                         style="font-size: 1rem; color:#303030;"></i>
-                        </button>
+                        <button type="button" class="btn btn-outline-danger btn-sm" data-bs-toggle="modal" data-bs-target="#A{{ $listas->cpf}}-{{ $listas->idf }}"class="btn btn-outline-danger btn-sm"><i class="bi-trash" style="font-size: 1rem; color:#303030;"></i></button>
 
 
                         <!-- <a href="/excluir-funcionario/{{$listas->idf}}"><button type="button" class="btn btn-outline-danger btn-sm" data-bs-toggle="modal fader" data-bs-target="#A{{ $listas->cpf}}-{{ $listas->idf }}"><i class="bi-trash" style="font-size: 1rem; color:#303030;"></i></button></a>
@@ -120,8 +80,7 @@
 
                     <!-- Modal -->
                     <div>
-                        <div class="modal fade" id="A{{ $listas->cpf}}-{{ $listas->idf }}" tabindex="-1"
-                             aria-labelledby="exampleModalLabel" aria-hidden="true">
+                        <div class="modal fade" id="A{{ $listas->cpf}}-{{ $listas->idf }}" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true" >
                             <div class="modal-dialog">
                                 <div class="modal-content">
                                     <div class="modal-header">
@@ -136,11 +95,8 @@
                                         </p>
                                     </div>
                                     <div class="modal-footer">
-                                        <button type="button" class="btn btn-danger" data-bs-dismiss="modal">Cancelar
-                                        </button>
-                                        <a href="/excluir-funcionario/{{$listas->idf}}">
-                                            <button type="button" class="btn btn-primary">Confirmar</button>
-                                        </a>
+                                        <button type="button" class="btn btn-danger" data-bs-dismiss="modal">Cancelar</button>
+                                        <a href="/excluir-funcionario/{{$listas->idf}}"><button type="button" class="btn btn-primary">Confirmar</button></a>
                                     </div>
                                 </div>
                             </div>
@@ -165,13 +121,9 @@
     </div>
     </div>
     </div>
-    <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.8/dist/umd/popper.min.js"
-            integrity="sha384-I7E8VVD/ismYTF4hNIPjVp/Zjvgyol6VFvRkX/vR+Vc4jQkC+hVqc2pM8ODewa9r"
-            crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.min.js"
-            integrity="sha384-BBtl+eGJRgqQAUMxJ7pMwbEyER4l1g+O15P+16Ep7Q9Q+zqX6gSbd85u4mG4QzX+"
-            crossorigin="anonymous"></script>
 @endsection
 
 
+@section('footerScript')
 
+@endsection
